+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
+## [Unreleased]
+### Changed
+- `JSONExpect`: improved error messages
+- `JSONExpect`: all tests now strongly-typed
+
 ## [0.2] - 2020-04-13
 ### Changed
 - `JSONExpect`: complete rewrite; much simpler and faster